refactor(Standalone): Seclude standalone utils
