+++ lib/plugins/standalone/index.js
@@ -2,17 +2,17 @@
 
 const fs = require('fs');
 const os = require('os');
+const path = require('path');
 const streamPromise = require('stream-promise');
 const fse = require('fs-extra');
 const fetch = require('node-fetch');
-const isInChina = require('@serverless/utils/is-in-china');
 const currentVersion = require('../../../package').version;
 const isStandaloneExecutable =
   require('../../utils/isStandaloneExecutable') && process.platform !== 'win32';
+const standaloneUtils = require('../../utils/standalone');
 
-const BINARIES_DIR_PATH = `${os.homedir()}/.serverless/bin`;
-const BINARY_TMP_PATH = `${BINARIES_DIR_PATH}/serverless-tmp`;
-const BINARY_PATH = `${BINARIES_DIR_PATH}/serverless`;
+const BINARY_TMP_PATH = os.tmpdir();
+const BINARY_PATH = standaloneUtils.path;
 
 module.exports = class Standalone {
   constructor(serverless, cliOptions) {
@@ -48,20 +48,7 @@ module.exports = class Standalone {
   }
 
   async upgrade() {
-    const latestTagResponse = await fetch(
-      isInChina
-        ? 'https://sls-standalone-1300963013.cos.ap-shanghai.myqcloud.com/latest-tag'
-        : 'https://api.github.com/repos/serverless/serverless/releases/latest'
-    );
-    if (!latestTagResponse.ok) {
-      throw new this.serverless.classes.Error(
-        'Sorry unable to `upgrade` at this point ' +
-          `(server rejected request with ${latestTagResponse.status})`
-      );
-    }
-    const tagName = isInChina
-      ? await latestTagResponse.text()
-      : (await latestTagResponse.json()).tag_name;
+    const tagName = await standaloneUtils.resolveLatestTag();
     const latestVersion = tagName.slice(1);
     if (latestVersion === currentVersion) {
       this.serverless.cli.log('Already at latest version');
@@ -82,31 +69,8 @@ module.exports = class Standalone {
         ].join('\n')
       );
     }
-    const platform = (() => {
-      switch (process.platform) {
-        case 'darwin':
-          return 'macos';
-        default:
-          return process.platform;
-      }
-    })();
-    const arch = (() => {
-      switch (process.arch) {
-        case 'x32':
-          return 'x86';
-        case 'arm':
-        case 'arm64':
-          return 'armv6';
-        default:
-          return process.arch;
-      }
-    })();
     this.serverless.cli.log('Downloading new version...');
-    const executableUrl = isInChina
-      ? `https://sls-standalone-1300963013.cos.ap-shanghai.myqcloud.com/${tagName}/` +
-        `serverless-${platform}-${arch}`
-      : `https://github.com/serverless/serverless/releases/download/${tagName}/` +
-        `serverless-${platform}-${arch}`;
+    const executableUrl = standaloneUtils.resolveUrl(tagName);
     const standaloneResponse = await fetch(executableUrl);
     if (!standaloneResponse.ok) {
       throw new this.serverless.classes.Error(
@@ -121,7 +85,7 @@ module.exports = class Standalone {
   }
 
   async uninstall() {
-    await fse.remove(BINARIES_DIR_PATH);
+    await fse.remove(path.dirname(BINARY_PATH));
     this.serverless.cli.log('Uninstalled');
   }
